+++ crates/parol/src/bin/parol/arguments.rs
@@ -79,6 +79,11 @@ pub(crate) struct CliArgs {
     #[arg(long, value_enum)]
     pub inner_attributes: Vec<InnerAttributes>,
 
+    /// Disbales the error recovery mechanism in the generated parser
+    /// This has currently no effect on LR parsers, because error recovery is not available yet.
+    #[arg(long)]
+    pub disable_recovery: bool,
+
     #[command(subcommand)]
     pub subcommand: Option<super::tools::ToolsSubcommands>,
 }