+++ crates/parol/CHANGELOG.md
@@ -27,11 +27,10 @@ new minor version numbers. Please note that this is no version handling covered 
 of regexes created with the help of the `regex-automata` crate. Please have a look at
 [parol_runtime's CHANGELOG](../parol_runtime/CHANGELOG.md).
 
-- The version 1 will be supported and updated regularly on branch `release1.0`, so you aren't forced
-to switch to version 2 any time soon.
-
 - Refactor some grammar tools to enforce LL grammar type
 
+- Error recovery on generated LL(k) parsers can now be disabled, via Builder API (`disable_recovery`)
+and via command line argument (`--disable_recovery`)
 
 ## 1.0.0 - 2024-09-02